requirements.txt
certifi==2020.6.20
cffi==1.14.0
click==7.1.2
cryptography==2.9.2
decorator==4.4.2
future==0.18.2
joblib==0.15.1
nltk==3.5
numpy==1.19.0
pycparser==2.20
python-telegram-bot==12.7
regex==2020.6.8
scikit-learn==0.23.1
scipy==1.5.0
six==1.15.0
threadpoolctl==2.1.0
tornado==6.0.4
tqdm==4.46.1
